# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/image/metadatagenerator": {
      "post": {
        "summary": "Metadata Generator",
        "description": "Generate metadata from the image.\n\nArgs:\n    metadata_request_data: image\n\nReturns:\n    The generated metadata",
        "operationId": "metadata_generator_api_v1_image_metadatagenerator_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetaDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GenerateOptions": {
        "properties": {
          "image_description": {
            "type": "boolean",
            "title": "Image Description"
          },
          "image_title": {
            "type": "boolean",
            "title": "Image Title"
          },
          "image_sentiment": {
            "type": "boolean",
            "title": "Image Sentiment"
          }
        },
        "type": "object",
        "required": [
          "image_description",
          "image_title",
          "image_sentiment"
        ],
        "title": "GenerateOptions",
        "description": "Generate options for the image metadata generation endpoint."
      },
      "GeneratedMetadata": {
        "properties": {
          "image_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Description"
          },
          "image_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Title"
          },
          "image_sentiment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Sentiment"
          }
        },
        "type": "object",
        "required": [
          "image_description",
          "image_title",
          "image_sentiment"
        ],
        "title": "GeneratedMetadata",
        "description": "Generated metadata for the image."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MetaDataRequest": {
        "properties": {
          "image_base64": {
            "type": "string",
            "title": "Image Base64"
          },
          "generate_options": {
            "$ref": "#/components/schemas/GenerateOptions"
          }
        },
        "type": "object",
        "required": [
          "image_base64",
          "generate_options"
        ],
        "title": "MetaDataRequest",
        "description": "Request body for the image metadata generation endpoint."
      },
      "MetaDataResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "data": {
            "$ref": "#/components/schemas/GeneratedMetadata"
          }
        },
        "type": "object",
        "required": [
          "status",
          "data"
        ],
        "title": "MetaDataResponse",
        "description": "Request body for the image metadata generation endpoint."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}